Remove unnecessary cast

+++ pulsar-broker/src/main/java/org/apache/pulsar/broker/service/Producer.java
@@ -265,7 +265,7 @@ public class Producer {
             }
         }
 
-        startPublishOperation((int) batchSize, msgSize);
+        startPublishOperation(batchSize, msgSize);
         return true;
     }
